requirements.txt
cryptography>=3.4
liboqs-python>=0.3.0
cbor2>=5.4.0